Always attempt a graceful container stop/remove
When paunch removes a container, like applying a container's config
changes, it does that not gracefully. That poses an issue for some
cases, like stateful services. To mitigate that, attempt stopping the
container prior to removing it. And use the defined stop signal and
timeout settings for such a container, when stopping it. If there is no
signal/timeout in the config, fallback to deaults of SIGTERM and a 10s.

Related-bug: #1810690
Change-Id: I7c72ea055984bb5ff3ea7cea019693d61245212f
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>